requirements.txt
# tested with python 3.6.9
pymongo==3.10.1
numpy==1.18.3
requests==2.23.0
Pillow==7.1.1
scipy==1.4.1
matplotlib==3.3.4
plotly
pandas
moviepy
mediapipe
opencv-python
pascal-voc-writer
detecto